WIP: structure of constant import changed, notification constants renamed
