+++ src/components/Login.vue
@@ -32,7 +32,7 @@
 
 <script>
 
-import { AUTH } from '../constants/stores';
+import { AUTH, NOTIFICATION_TYPES } from '../constants';
 
 export default {
   name: 'Login',
@@ -48,12 +48,12 @@ export default {
       this.$store.dispatch(AUTH.ACTIONS.AUTH_REQUEST, { login, password })
         .then(() => this.$router.push(this.$route.query.redirect || '/'))
         .then(() => this.$notify({
-          type: 'success',
+          type: NOTIFICATION_TYPES.SUCCESS,
           text: 'Login successful',
         }))
         .catch((err) => {
           this.$notify({
-            type: 'error',
+            type: NOTIFICATION_TYPES.ERROR,
             text: err,
           });
         });
