+++ src/constants/index.js
@@ -1,0 +1,9 @@
+export {
+  STATISTICS_PATHS, STATISTICS_ROUTE_NAMES, GENERAL_PATHS, GENERAL_ROUTE_NAMES,
+} from './routing';
+
+export {
+  RATES, ACCOUNTS, CDRS, AUTH,
+} from './stores';
+
+export { NOTIFICATION_TYPES } from './common';
