+++ src/store/modules/accounts.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line
 import { jsonApi } from '../../api';
 import { RESOURCES } from '../../static/constants/api';
-import { ACCOUNTS } from '../../constants/stores';
+import { ACCOUNTS } from '../../constants';
 
 const state = {
   accounts: {},
